+++ py.sdk.todos/CHANGELOG.md
@@ -5,7 +5,7 @@ Most notable changes to this project will be documented in this file.
 Updates follow a [conventional commit][commits] style and the project is
 versioned with [calendar versioning][calver].
 
-- build: package scripts as commands without needing installation 2025-12-18
+- feat: package scripts as commands to manage tasks in lists 2026-02-08
 
 [calver]: https://calver.org
 [commits]: https://www.conventionalcommits.org/en/v1.0.0/